LibBladeRF/Sync.hs: bladeRFSyncTx already has sample length info
Since we use a ByteString passed to bladeRFSyncTx we do not need
to explicitly pass the number of samples as we have that
information encoded in the ByteString itself. Hence, let us make
use of this.

Note here that: The constant 4 used here is because ByteString
is a sequence of bytes, there are 2 bytes in a signed 16-bit
integer, and there are 2 signed 16-bit integers in a sample.

Signed-off-by: Edward O'Callaghan <eocallaghan@alterapraxis.com>